requirements.txt
# LangChain & LangGraph
langchain==0.3.7
langchain-openai==0.2.5
langchain-community==0.3.5
langgraph==0.2.45
langchain-postgres==0.0.12

# Vector Store
chromadb==0.5.20

# Observability & Evaluation
langfuse==3.10.0
langchain-langfuse==2.0.3
ragas==0.2.5

# Database
psycopg2-binary==2.9.10

# UI
streamlit==1.40.1

# Data Processing
pydantic==2.9.2
python-dotenv==1.0.1

# Additional Dependencies
numpy==1.26.4
pandas==2.2.3
tiktoken==0.8.0
nltk
